requirements.txt
# PDF extraction
pypdf>=4.0.0

# Embeddings and vector search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# LLM (OpenAI, Ollama, Groq, OpenRouter) and env
openai>=1.0.0
python-dotenv>=1.0.0

# API and UI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.28.0
watchdog>=3.0.0

# Eval: BLEU (nltk), ROUGE (rouge-score)
nltk>=3.8.0
rouge-score>=0.1.2

# Development
pytest>=7.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
